Begin rewrite
When designing the project architecture I didn't understand the usage of
hooks, and tried to embroil all logic into the TreeStart hook.

This has caused several problems, in particular the bot not defending
itself when attacked. As a result, I think the best course of action is
to redesign all the behaviour.