Improve combat handling
Add full support for Kupo, which seems to have some sort of bug
preventing it from moving unless its pull behaviour is also called. It
may be me not fully understanding how to call it, but other combat
routines don't seem to have this issue.

Added check to see if the current Poi mob is a part of a FATE that is
gone, so that we can clear it.

Added check to see if the current Poi stored in Tarot is null before
reassigning, so we don't assign a null Poi and cause an exception.

Moved movement handling in the FATE handler to its own task.