Fix several combat Poi bugs
Fix thrashing of Poi if FATE is no longer active and current target is a
part of that FATE.

Fix null reference exception if combat Poi's BattleCharacter no longer
exists. This could occur if the combat routine was resting for an
extended period of time after killing the mob, and the mob's corpse
despawns before we exit the resting period.