Fill in missing data
Fill in missing data for two collection FATEs.

Setting 'Obey their Thirst' as unsupported for now, as I could not seem
to get the FATE to spawn and thus, cannot get information about the turn
in NPC. It also has some navigation issues that may not be fixable with
waypoints.

Also did some code clean up.