Implement waiting for boss FATEs
Bot will now wait for a set amount of progress to be met before engaging
in boss and mega-boss FATEs. This can be configured by the user, and
also allows the user to choose whether or not to pick a FATE that is
below the minimum progress.

Refactored FateHandler to be more readable.