Implement wait random time before moving to FATE
Fixed an issue with combat where if the bot navigated out of the FATE
area when trying to reach a FATE mob, it would get stuck in a loop of
going back into the FATE area, then leaving it again. The bot will now
only go back into the FATE area if it's already in combat with the mob
it's trying to attack.