Numerous changes
Removed FATE chain on failure behaviour, it's not feasible to check
whether a FATE succeeded or failed. Easier to only wait for a FATE if it
succeeded.

Added a check to see if the next FATE in a chain is on the user
blacklist, or if its type is disabled. If either of those, don't bother
waiting.

Added a safety check for moving to FATEs, so that the position will
update should the FATE location move.

Cleaned up the return to aetheryte wait behaviour.

Removed disabling of ExBuddy's flight navigator. If we use our own
internal AStar navigator for flight meshed zones we don't have to worry
about it overriding RebornBuddy's navigation provider.