Bug fixes, flight logic improvements
Fixed a bug where if the bot stopped while loading a flight mesh, the
file stream wouldn't be closed automatically. The file stream is now
always closed when the bot stops.

Fixed a bug where the bot would occasionally get stuck when trying to
land. We now just try to land in any location, rather than polling the
CanLand task.

We now send a raycast to the second node when beginning flight to see if
the first can be skipped, rather than just assuming it can be.