Fix several flight navigation issues
Oracle will no longer sometimes get stuck trying to take off (was being
caused by an attempt to move to a node beneath us).

Revamped the landing logic to be more stable when handling difficult
locations.

We now prefer landing further way from the centre of the FATE, should
help survivability a bit.

Added a preferred mob for the FATE 'Saved the Nests for Last'.