Landing logic improvements.
Reduce the range of variance for the flight path post-processing.

Flag the Darkscale FATEs as mega-bosses. Regardless of what the icon says, these FATEs are much closer to that of a mega-boss FATE than a defence FATE.

Add new landing radius property to the FATE database, indicating if a FATE's landing areas are limited.

Cleaned up the settings file.