Implement true level checking.
The ClassLevel property of a LocalPlayer object in RebornBuddy will show a synced level if we're level synced. This appears to be new and was causing issues with Oracle. To solve, GetTrueLevel has been implemented which will search the Levels property of a LocalPlayer for our current class instead.

Added a timeout if the mob we're targetting doesn't take damage for 15 seconds.