For closest FATEs, distance from aetherytes is now taken into account.
Option to turn this on and off added. Probably need to clarify this setting in the UI as it's not intuitive.

Fixed an issue with the Little Ala Mhigo aetheryte location not being taken into account when searching for lower distances. Suspected cause is the location of the aetheryte being deemed unnavigable by the navigation server, so have overridden the default location with our own.