Add option to override the combat routine's rest behaviour.
Some routines, such as Ultima, have pretty obnoxious rest behaviours that lock the bot while it waits for an arbitrary amount of HP/MP. We should only rest if we have a combat Poi and our HP/MP are below the minimum level the user sets.

Improve follow behaviour for escort FATEs: Oracle should navigate back to the FATE centre more reliably if it falls out of the radius.